+++ src/rofi_rbw/notifier.py
@@ -1,0 +1,14 @@
+from subprocess import run
+
+
+class Notifier:
+    def notify(self, message: str) -> None:
+        run(
+            [
+                "notify-send",
+                "-u",
+                "normal",
+                "rofi-rbw",
+                message,
+            ],
+        )